refactor: devpool-claims to rewards rename
