+++ static/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <link href="out/styles/devpool-claims/devpool-claims.css" rel="stylesheet" />
+    <link href="out/styles/rewards/rewards.css" rel="stylesheet" />
     <link rel="canonical" href="https://pay.ubq.fi" />
     <meta charset="utf-8" />
     <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
@@ -157,6 +157,6 @@
     </main>
 
     <ul class="notifications"></ul>
-    <script src="out/scripts/devpool-claims/index.js" type="application/javascript"></script>
+    <script src="out/scripts/rewards/index.js" type="application/javascript"></script>
   </body>
 </html>
