+++ static/scripts/rewards/abis/permit2Abi.ts
