+++ static/scripts/rewards/cirip/query-graph.ts
