+++ static/scripts/rewards/constants.ts
