+++ static/scripts/rewards/get-contract.ts
