+++ static/scripts/rewards/index.ts
