+++ static/scripts/rewards/invalidate-component.ts
