+++ static/scripts/rewards/render-transaction/index.ts
