+++ static/scripts/rewards/render-transaction/insert-table-data.ts
