+++ static/scripts/rewards/render-transaction/render-token-symbol.ts
