+++ static/scripts/rewards/render-transaction/render-transaction.ts
