+++ static/scripts/rewards/render-transaction/set-claim-message.ts
