+++ static/scripts/rewards/render-transaction/tx-type.ts
