+++ static/scripts/rewards/the-grid.ts
