+++ static/scripts/rewards/toaster.ts
