+++ static/scripts/rewards/web3/connect-wallet.ts
