+++ static/scripts/rewards/web3/invalidate-nonce.ts
