+++ static/scripts/rewards/web3/pay.ts
