+++ static/scripts/rewards/web3/switch-network.ts
