+++ static/styles/audit-report/audit.css
@@ -1,4 +1,4 @@
-@import url("../devpool-claims/pay.css");
+@import url("../rewards/pay.css");
 @import url("../proxima.css");
 @import url("checkbox.css");
 @import url("../fa.css");
