+++ static/styles/rewards/background.css
