+++ static/styles/rewards/claim-table.css
