+++ static/styles/rewards/light-mode.css
