+++ static/styles/rewards/pay.css
